refactor: extract PR body rendering into rebasebot.pr_body
Move build_pr_body and its display-limit constants out of bot.py into
a dedicated pure module, matching the prow.py and rebase_summary.py
pattern. Relocate TestBuildPrBody to tests/test_pr_body.py.
