+++ rebasebot/bot.py
@@ -37,6 +37,7 @@ from github3.repos.repo import Repository
 from rebasebot import lifecycle_hooks
 from rebasebot.github import GithubAppProvider, GitHubBranch
 from rebasebot.lifecycle_hooks import LifecycleHookScriptException
+from rebasebot.pr_body import build_pr_body
 from rebasebot.prow import ProwJobContext
 from rebasebot.rebase_summary import ArtPrInfo, ContentLossWarning, DroppedCommit, RebaseSummary
 from rebasebot.slack import SlackNotifier
@@ -59,8 +60,6 @@ MERGE_TMP_BRANCH = "merge-tmp"
 _COMMIT_LOG_FORMAT = "--pretty=format:%H || %s || %aE"
 _MERGE_COMMIT_PARENT_COUNT = 2
 _LOST_LINE_LOG_LIMIT = 10
-_PR_BODY_LOST_LINE_LIMIT = 20
-_DROPPED_COMMITS_DISPLAY_LIMIT = 20
 _GO_MODULES_CARRY_COMMIT_MESSAGE = "UPSTREAM: <carry>: Updating and vendoring go modules after an upstream rebase"
 
 
@@ -810,82 +809,6 @@ def _is_pr_available(dest_repo: Repository, dest: GitHubBranch, rebase: GitHubBr
     return None, False
 
 
-def _build_pr_body(
-    summary: RebaseSummary,
-    source: GitHubBranch,
-    dest: GitHubBranch,
-    prow_job: ProwJobContext,
-) -> str:
-    """Render the Markdown body for a rebase pull request."""
-    sections: list[str] = [
-        "This is an automated rebase PR generated by RebaseBot.",
-    ]
-
-    if summary.upstream_commit_count == 1:
-        commit_line = "1 new upstream commit"
-    else:
-        commit_line = f"{summary.upstream_commit_count} new upstream commits"
-
-    summary_lines = [
-        f"- **Source**: `{source.url}:{source.branch}`",
-        f"- **Destination**: `{dest.url}:{dest.branch}`",
-        f"- **{commit_line}**",
-    ]
-    sections.append("## Summary\n" + "\n".join(summary_lines))
-
-    if summary.dropped_commits:
-        dropped_lines = [
-            f"- `{commit.sha[:7]}` {commit.message} ({commit.reason})"
-            for commit in summary.dropped_commits[:_DROPPED_COMMITS_DISPLAY_LIMIT]
-        ]
-        remaining = len(summary.dropped_commits) - _DROPPED_COMMITS_DISPLAY_LIMIT
-        if remaining > 0:
-            dropped_lines.append(f"- ... and {remaining} more")
-        sections.append("## Dropped downstream commits\n" + "\n".join(dropped_lines))
-
-    if summary.art_pr is not None:
-        sections.append(
-            "## ART pull request cherry-picked\n\n"
-            f"[#{summary.art_pr.number} {summary.art_pr.title}]({summary.art_pr.url})"
-        )
-
-    if summary.content_loss_warnings:
-        warnings_by_commit: dict[tuple[str, str], list[ContentLossWarning]] = {}
-        commit_order: list[tuple[str, str]] = []
-        for warning in summary.content_loss_warnings:
-            key = (warning.sha, warning.message)
-            if key not in warnings_by_commit:
-                warnings_by_commit[key] = []
-                commit_order.append(key)
-            warnings_by_commit[key].append(warning)
-
-        details_blocks: list[str] = []
-        for key in commit_order:
-            sha, message = key
-            file_sections: list[str] = []
-            for warning in warnings_by_commit[key]:
-                displayed_lines = warning.lost_lines[:_PR_BODY_LOST_LINE_LIMIT]
-                code_block = "\n".join(line.strip() for line in displayed_lines)
-                file_section = f"**{warning.file}**\n```\n{code_block}\n```"
-                remaining = len(warning.lost_lines) - _PR_BODY_LOST_LINE_LIMIT
-                if remaining > 0:
-                    file_section += f"\n... and {remaining} more lines"
-                file_sections.append(file_section)
-
-            details_blocks.append(
-                f"<details>\n<summary>`{sha[:7]}` {message}</summary>\n\n"
-                + "\n\n".join(file_sections)
-                + "\n\n</details>"
-            )
-
-        sections.append("## ⚠️ Possible upstream content loss\n\n" + "\n\n".join(details_blocks))
-
-    if prow_job.log_url is not None:
-        sections.append(f"## Logs\n\n[View job log]({prow_job.log_url})")
-
-    return "\n\n".join(sections)
-
-
 def _create_pr(
     *,
     gh_app: github3.GitHub,
@@ -917,7 +840,7 @@ def _create_pr(
         f"https://api.github.com/repos/{dest.ns}/{dest.name}/pulls",
         data={
             "title": title,
-            "body": _build_pr_body(summary, source, dest, prow_job),
+            "body": build_pr_body(summary, source, dest, prow_job),
             "head": rebase.branch,
             "head_repo": f"{rebase.ns}/{rebase.name}",
             "base": dest.branch,
@@ -1082,7 +1005,7 @@ def _update_pr_body(
     prow_job: ProwJobContext,
 ) -> None:
     """Regenerate and overwrite the pull request body on every push."""
-    body = _build_pr_body(summary, source, dest, prow_job)
+    body = build_pr_body(summary, source, dest, prow_job)
     logging.info("Updating pull request body")
     if not pull_req.update(body=body):
         raise PullRequestUpdateException(f"Error updating body for pull request: {pull_req.html_url}")
