+++ rebasebot/pr_body.py
@@ -1,0 +1,99 @@
+#    Copyright 2022 Red Hat, Inc.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+"""Markdown rendering for rebase pull request bodies."""
+
+from __future__ import annotations
+
+from rebasebot.github import GitHubBranch
+from rebasebot.prow import ProwJobContext
+from rebasebot.rebase_summary import ContentLossWarning, RebaseSummary
+
+PR_BODY_LOST_LINE_LIMIT = 20
+DROPPED_COMMITS_DISPLAY_LIMIT = 20
+
+
+def build_pr_body(
+    summary: RebaseSummary,
+    source: GitHubBranch,
+    dest: GitHubBranch,
+    prow_job: ProwJobContext,
+) -> str:
+    """Render the Markdown body for a rebase pull request."""
+    sections: list[str] = [
+        "This is an automated rebase PR generated by RebaseBot.",
+    ]
+
+    if summary.upstream_commit_count == 1:
+        commit_line = "1 new upstream commit"
+    else:
+        commit_line = f"{summary.upstream_commit_count} new upstream commits"
+
+    summary_lines = [
+        f"- **Source**: `{source.url}:{source.branch}`",
+        f"- **Destination**: `{dest.url}:{dest.branch}`",
+        f"- **{commit_line}**",
+    ]
+    sections.append("## Summary\n" + "\n".join(summary_lines))
+
+    if summary.dropped_commits:
+        dropped_lines = [
+            f"- `{commit.sha[:7]}` {commit.message} ({commit.reason})"
+            for commit in summary.dropped_commits[:DROPPED_COMMITS_DISPLAY_LIMIT]
+        ]
+        remaining = len(summary.dropped_commits) - DROPPED_COMMITS_DISPLAY_LIMIT
+        if remaining > 0:
+            dropped_lines.append(f"- ... and {remaining} more")
+        sections.append("## Dropped downstream commits\n" + "\n".join(dropped_lines))
+
+    if summary.art_pr is not None:
+        sections.append(
+            "## ART pull request cherry-picked\n\n"
+            f"[#{summary.art_pr.number} {summary.art_pr.title}]({summary.art_pr.url})"
+        )
+
+    if summary.content_loss_warnings:
+        warnings_by_commit: dict[tuple[str, str], list[ContentLossWarning]] = {}
+        commit_order: list[tuple[str, str]] = []
+        for warning in summary.content_loss_warnings:
+            key = (warning.sha, warning.message)
+            if key not in warnings_by_commit:
+                warnings_by_commit[key] = []
+                commit_order.append(key)
+            warnings_by_commit[key].append(warning)
+
+        details_blocks: list[str] = []
+        for key in commit_order:
+            sha, message = key
+            file_sections: list[str] = []
+            for warning in warnings_by_commit[key]:
+                displayed_lines = warning.lost_lines[:PR_BODY_LOST_LINE_LIMIT]
+                code_block = "\n".join(line.strip() for line in displayed_lines)
+                file_section = f"**{warning.file}**\n```\n{code_block}\n```"
+                remaining = len(warning.lost_lines) - PR_BODY_LOST_LINE_LIMIT
+                if remaining > 0:
+                    file_section += f"\n... and {remaining} more lines"
+                file_sections.append(file_section)
+
+            details_blocks.append(
+                f"<details>\n<summary>`{sha[:7]}` {message}</summary>\n\n"
+                + "\n\n".join(file_sections)
+                + "\n\n</details>"
+            )
+
+        sections.append("## ⚠️ Possible upstream content loss\n\n" + "\n\n".join(details_blocks))
+
+    if prow_job.log_url is not None:
+        sections.append(f"## Logs\n\n[View job log]({prow_job.log_url})")
+
+    return "\n\n".join(sections)
