+++ tests/test_bot.py
@@ -22,7 +22,6 @@ from rebasebot.bot import (
     PullRequestUpdateException,
     RepoException,
     _add_to_rebase,
-    _build_pr_body,
     _is_pr_available,
     _is_pr_required,
     _report_result,
@@ -31,8 +30,9 @@ from rebasebot.bot import (
     notify_slack_error,
 )
 from rebasebot.github import GitHubBranch
+from rebasebot.pr_body import build_pr_body
 from rebasebot.prow import ProwJobContext
-from rebasebot.rebase_summary import ArtPrInfo, ContentLossWarning, DroppedCommit, RebaseSummary
+from rebasebot.rebase_summary import RebaseSummary
 from rebasebot.slack import _build_slack_blocks
 
 from .conftest import CommitBuilder
@@ -584,301 +584,6 @@ class TestUpdatePrTitle:
         )
 
 
-class TestBuildPrBody:
-    def test_with_log_url(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(upstream_commit_count=3)
-        prow_job = ProwJobContext(
-            job_name="periodic-openshift-release-rebasebot",
-            job_type="periodic",
-            build_id="12345",
-        )
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert body == (
-            "This is an automated rebase PR generated by RebaseBot.\n\n"
-            "## Summary\n"
-            "- **Source**: `https://github.com/upstream/repo:main`\n"
-            "- **Destination**: `https://github.com/downstream/repo:release`\n"
-            "- **3 new upstream commits**\n\n"
-            "## Logs\n\n"
-            "[View job log](https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/"
-            "periodic-openshift-release-rebasebot/12345)"
-        )
-
-    def test_without_log_url(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(upstream_commit_count=1)
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert body == (
-            "This is an automated rebase PR generated by RebaseBot.\n\n"
-            "## Summary\n"
-            "- **Source**: `https://github.com/upstream/repo:main`\n"
-            "- **Destination**: `https://github.com/downstream/repo:release`\n"
-            "- **1 new upstream commit**"
-        )
-        assert "## Logs" not in body
-
-    def test_with_dropped_commits(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(
-            upstream_commit_count=2,
-            dropped_commits=[
-                DroppedCommit(
-                    sha="abcdef1234567890",
-                    message="UPSTREAM: <carry>: excluded patch",
-                    reason="explicitly excluded via --exclude-commits",
-                ),
-                DroppedCommit(
-                    sha="1234567890abcdef",
-                    message="untagged commit",
-                    reason="dropped by tag policy",
-                ),
-            ],
-        )
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert "## Dropped downstream commits" in body
-        assert "- `abcdef1` UPSTREAM: <carry>: excluded patch (explicitly excluded via --exclude-commits)" in body
-        assert "- `1234567` untagged commit (dropped by tag policy)" in body
-        assert "## ART pull request cherry-picked" not in body
-
-    def test_without_dropped_commits(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(upstream_commit_count=0)
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert "## Dropped downstream commits" not in body
-
-    def test_dropped_commits_truncation(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        dropped_commits = [
-            DroppedCommit(
-                sha=f"{index:040x}",
-                message=f"commit {index}",
-                reason="dropped by tag policy",
-            )
-            for index in range(25)
-        ]
-        summary = RebaseSummary(upstream_commit_count=25, dropped_commits=dropped_commits)
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        dropped_section = body.split("## Dropped downstream commits\n", 1)[1]
-        assert dropped_section.count("- `") == 20
-        assert "- ... and 5 more" in dropped_section
-        assert "commit 19" in dropped_section
-        assert "commit 20" not in dropped_section
-
-    def test_with_art_pr(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(
-            upstream_commit_count=1,
-            art_pr=ArtPrInfo(
-                number=42,
-                title="Update build image to be consistent with ART",
-                url="https://github.com/downstream/repo/pull/42",
-            ),
-        )
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert "## ART pull request cherry-picked" in body
-        assert (
-            "[#42 Update build image to be consistent with ART](https://github.com/downstream/repo/pull/42)"
-        ) in body
-        assert "## Dropped downstream commits" not in body
-
-    def test_with_dropped_commits_and_art_pr(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(
-            upstream_commit_count=2,
-            dropped_commits=[
-                DroppedCommit(
-                    sha="abcdef1234567890",
-                    message="untagged commit",
-                    reason="dropped by tag policy",
-                )
-            ],
-            art_pr=ArtPrInfo(
-                number=7,
-                title="Update build image to be consistent with ART",
-                url="https://github.com/downstream/repo/pull/7",
-            ),
-        )
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert "## Dropped downstream commits" in body
-        assert "## ART pull request cherry-picked" in body
-        assert body.index("## Dropped downstream commits") < body.index("## ART pull request cherry-picked")
-
-    def test_with_content_loss_warnings(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(
-            upstream_commit_count=1,
-            content_loss_warnings=[
-                ContentLossWarning(
-                    sha="abcdef1234567890",
-                    message="UPSTREAM: <carry>: add snapshot timeout",
-                    file="test.go",
-                    lost_lines=[
-                        '\tebsKmsKeyIDKey = "ebsKmsKeyId"',
-                        "\tebsKmsKeyId string",
-                    ],
-                )
-            ],
-        )
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert "## ⚠️ Possible upstream content loss" in body
-        assert "<details>" in body
-        assert "<summary>`abcdef1` UPSTREAM: <carry>: add snapshot timeout</summary>" in body
-        assert "**test.go**" in body
-        assert 'ebsKmsKeyIDKey = "ebsKmsKeyId"' in body
-        assert "## Dropped downstream commits" not in body
-        assert "## ART pull request cherry-picked" not in body
-
-    def test_without_content_loss_warnings(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(upstream_commit_count=1)
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert "## ⚠️ Possible upstream content loss" not in body
-
-    def test_content_loss_line_truncation(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        lost_lines = [f"lost line {index}" for index in range(25)]
-        summary = RebaseSummary(
-            upstream_commit_count=1,
-            content_loss_warnings=[
-                ContentLossWarning(
-                    sha="abcdef1234567890",
-                    message="conflicting commit",
-                    file="large.go",
-                    lost_lines=lost_lines,
-                )
-            ],
-        )
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert "lost line 0" in body
-        assert "lost line 19" in body
-        assert "lost line 20" not in body
-        assert "... and 5 more lines" in body
-
-    def test_with_dropped_commits_art_pr_and_content_loss(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(
-            upstream_commit_count=2,
-            dropped_commits=[
-                DroppedCommit(
-                    sha="abcdef1234567890",
-                    message="untagged commit",
-                    reason="dropped by tag policy",
-                )
-            ],
-            art_pr=ArtPrInfo(
-                number=7,
-                title="Update build image to be consistent with ART",
-                url="https://github.com/downstream/repo/pull/7",
-            ),
-            content_loss_warnings=[
-                ContentLossWarning(
-                    sha="fedcba0987654321",
-                    message="UPSTREAM: <carry>: conflicting patch",
-                    file="test.go",
-                    lost_lines=["\tebsKmsKeyId string"],
-                )
-            ],
-        )
-        prow_job = ProwJobContext(job_name=None, job_type=None, build_id=None)
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        assert "## Dropped downstream commits" in body
-        assert "## ART pull request cherry-picked" in body
-        assert "## ⚠️ Possible upstream content loss" in body
-        assert body.index("## Dropped downstream commits") < body.index("## ART pull request cherry-picked")
-        assert body.index("## ART pull request cherry-picked") < body.index("## ⚠️ Possible upstream content loss")
-        assert "<details>" in body
-        assert "ebsKmsKeyId string" in body
-
-    def test_full_section_ordering_with_all_sections_and_log_url(self):
-        source = GitHubBranch(url="https://github.com/upstream/repo", ns="upstream", name="repo", branch="main")
-        dest = GitHubBranch(url="https://github.com/downstream/repo", ns="downstream", name="repo", branch="release")
-        summary = RebaseSummary(
-            upstream_commit_count=2,
-            dropped_commits=[
-                DroppedCommit(
-                    sha="abcdef1234567890",
-                    message="untagged commit",
-                    reason="dropped by tag policy",
-                )
-            ],
-            art_pr=ArtPrInfo(
-                number=7,
-                title="Update build image to be consistent with ART",
-                url="https://github.com/downstream/repo/pull/7",
-            ),
-            content_loss_warnings=[
-                ContentLossWarning(
-                    sha="fedcba0987654321",
-                    message="UPSTREAM: <carry>: conflicting patch",
-                    file="test.go",
-                    lost_lines=["\tebsKmsKeyId string"],
-                )
-            ],
-        )
-        prow_job = ProwJobContext(
-            job_name="periodic-openshift-release-rebasebot",
-            job_type="periodic",
-            build_id="12345",
-        )
-
-        body = _build_pr_body(summary, source, dest, prow_job)
-
-        section_headings = [
-            "## Summary",
-            "## Dropped downstream commits",
-            "## ART pull request cherry-picked",
-            "## ⚠️ Possible upstream content loss",
-            "## Logs",
-        ]
-        indices = [body.index(heading) for heading in section_headings]
-        assert indices == sorted(indices), "PR body sections must appear in order: " + " → ".join(
-            heading.removeprefix("## ").removeprefix("⚠️ ") for heading in section_headings
-        )
-
-
 class TestUpdatePrBody:
     def test_success(self):
         pull_req = MagicMock()
@@ -890,7 +595,7 @@ class TestUpdatePrBody:
 
         _update_pr_body(pull_req, summary, source, dest, prow_job)
 
-        pull_req.update.assert_called_once_with(body=_build_pr_body(summary, source, dest, prow_job))
+        pull_req.update.assert_called_once_with(body=build_pr_body(summary, source, dest, prow_job))
 
     def test_failure(self):
         pull_req = MagicMock()
